[UI] #267 Connect to PartyMemberDetail navigation
