+++ common/designsystem/src/main/res/values/strings.xml
@@ -17,6 +17,7 @@
     <string name="complete">확인</string>
     <string name="logout">로그아웃</string>
     <string name="clear">초기화</string>
+    <string name="participate">참여하기</string>
 
     <!-- app -->
     <string name="onboarding_first_banner_title">작품 정보뿐만 아니라 분실물까지!</string>
