+++ common/navigation/src/main/java/com/cmc/curtaincall/common/navigation/NavRouteLabel.kt
@@ -7,9 +7,11 @@ object NavRouteLabel {
     const val SIGNUP_TERMS = "signup_terms"
     const val SIGNUP_INPUT = "signup_input"
     const val WELCOME = "welcome"
+
     const val HOME = "home"
     const val HOME_GUIDE = "home_guide"
     const val HOME_REPORT = "home_report"
+
     const val SHOW_SEARCH = "show_search"
     const val SHOW_DETAIL = "show_detail"
     const val SHOW_REVIEW = "show_review"
@@ -17,12 +19,10 @@ object NavRouteLabel {
     const val SHOW_LOST_PROPERTY = "show_lost_property"
     const val SHOW_LOST_PROPERTY_DETAIL = "show_lost_property_detail"
     const val SHOW_LOST_PROPERTY_CREATE = "show_lost_property_create"
+
     const val PARTY_MEMBER = "party_member"
     const val PARTY_MEMBER_RECRUIT = "party_member_recruit"
-
-    const val PARTY_MEMBER_LIST = "party_member_list"
     const val PARTY_MEMBER_DETAIL = "party_member_detail"
-    const val PARTY_MEMBER_CREATE = "party_member_create"
 
     const val MYPAGE = "mypage"
     const val MYPAGE_PROFILE = "mypage_profile"
