+++ domain/src/main/java/com/cmc/curtaincall/domain/model/party/PartyDetailModel.kt
@@ -1,7 +1,6 @@
 package com.cmc.curtaincall.domain.model.party
 
 data class PartyDetailModel(
-    val category: String = "WATCHING",
     val content: String = "",
     val createdAt: String = "",
     val creatorId: Int = Int.MIN_VALUE,
