--- feature/partymember/src/main/java/com/cmc/curtaincall/feature/partymember/detail/PartyMemberDetailSideEffect.kt
@@ -1,8 +1,0 @@
-package com.cmc.curtaincall.feature.partymember.detail
-
-import com.cmc.curtaincall.core.base.BaseSideEffect
-
-sealed class PartyMemberDetailSideEffect : BaseSideEffect {
-    object SuccessDelete : PartyMemberDetailSideEffect()
-    object SuccessParticipation : PartyMemberDetailSideEffect()
-}
